Undo some changes
